Fixed low accuracy etc

+++ projectFor3Lab/src/main/resources/scripts/points/points_table_creation.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS points(
 	point_id SERIAL NOT NULL PRIMARY KEY,
-	xValue DECIMAL NOT NULL,
-	yValue DECIMAL NOT NULL,
+	xValue DOUBLE PRECISION NOT NULL,
+	yValue DOUBLE PRECISION NOT NULL,
 	function_id BIGINT NOT NULL,
 
 	CONSTRAINT function_id_fk FOREIGN KEY (function_id) REFERENCES math_functions (math_function_id)
